Create README.md
#Pong
This is a simple version of the game that kick started an industry.  I wanted to learn more about how to use the Phaser.js library in a real (albeit small) project and this is my goto exercise for learning new game libraries.  After all, if it makes Pong hard to do then we don't want to start Half Life 3 with it :)
I added some requirejs as well to manage the simple dependencies and for the learning opportunity.
I know the artwork is OK to use as it is all me - goggles on!  I believe the sound is OK to use as well but I grabbed these files years ago now so if I'm mistaken please let me know.
##Deploy
You may get away with just opening the index.html in your browser.  Really you need to start a web server - I was using the python -m SimpleHTTPServer in development until requirejs forced me to change to prevent caching.